+++ src/modes/modern.js
@@ -2,39 +2,41 @@ import { defaults, eventGroupByType, initialisersByGroup, initialiserParams } fr
 import extendWithKeyboardParams from '../utils/extendWithKeyboardParams';
 
 export default function () {
-	var simulant = function ( type, params = {} ) {
-		var event, group, Constructor, paramsList, paramName, i, extendedParams, isKeyboardEvent;
-
-		group = eventGroupByType[ type ];
+	function simulant ( type, params = {} ) {
+		let group = eventGroupByType[ type ];
+		let isKeyboardEvent;
 
 		if ( group === 'KeyboardEvent' ) {
 			group = 'Event'; // because you can't fake KeyboardEvents well in any browser
 			isKeyboardEvent = true;
 		}
 
-		Constructor = initialisersByGroup[ group ][0];
+		const initialiser = ( initialisersByGroup[ group ] || initialisersByGroup.Event );
 
-		extendedParams = {
+		const Constructor = initialiser[0] || window.Event;
+		const method = initialiser[1];
+
+		let extendedParams = {
 			bubbles: true, // TODO some events don't bubble?
 			cancelable: true
 		};
 
-		paramsList = initialiserParams[ initialisersByGroup[ group ][1] ];
-		i = ( paramsList ? paramsList.length : 0 );
+		const paramsList = initialiserParams[ method ];
+		let i = ( paramsList ? paramsList.length : 0 );
 
 		while ( i-- ) {
-			paramName = paramsList[i];
+			const paramName = paramsList[i];
 			extendedParams[ paramName ] = ( paramName in params ? params[ paramName ] : defaults[ paramName ] );
 		}
 
-		event = new Constructor( type, extendedParams );
+		let event = new Constructor( type, extendedParams );
 
 		if ( isKeyboardEvent ) {
 			extendWithKeyboardParams( event, params );
 		}
 
 		return event;
-	};
+	}
 
 	simulant.mode = 'modern';
 	return simulant;