+++ test/test.sh
@@ -1,2 +1,0 @@
-gobble build tmp -f
-phantomjs runner.js tmp/index.html