+++ CHANGELOG.md
@@ -1,14 +1,19 @@
 # changelog
 
-# 0.1.3
+## 0.1.4
+
+* handle `beforeunload` and similar events ([#6](https://github.com/Rich-Harris/simulant/issues/6))
+* internal: moved test suite to JSDOM
+
+## 0.1.3
 
 * `simulant.fire()` returns the event that was fired
 * Rewrote as ES6 modules, updated build process
 
-# 0.1.1, # 0.1.2
+## 0.1.1, # 0.1.2
 
 * fix manifest files
 
-# 0.1.0
+## 0.1.0
 
 * first release